requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
dnspython==1.16.0
fixerio==0.1.1
pymongo==3.12.0
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.10.0
six==1.16.0
tornado==6.1
tzlocal==2.1